Prognostic run: interpolate prescribed tendencies (#1563)
To do runs with physics prescribed from the SHiELD run, but at physics timesteps other than 15 minutes, we need to interpolate the prescribed Q values. This PR refactors the function time interpolation decorator we've been using in the nudging routine to be used in the tendency prescriber as well, since the context is similar (data is accessed via a timestep key and caching is advantageous). It also refactors some of the setup routines of the tendency prescriber to make testing easier.

Added public API:
- Added `reference_initial_time` and `reference_frequency_seconds` optional args to the `TendencyPrescriberConfig` (the same as are found in the nudging config). 

Refactored public API:
- Refactored function time interpolation routines into a new `runtime.interpolate` module.

Significant internal changes:
- Moved setup of `time_lookup_function` and the quantile limiter for the tendency prescriber into its own function, and pass the `time_lookup_function` function as an argument to the `TendencyPrescriber`.