Modify ClipConfig API to remove dimension, always clip last dim (#1687)
When we move to tfdatasets, we do not have access to dimension names. In practice we always want to clip the last dimension of an array. This PR modifies clipping configuration to remove the dimension name, and always assume we are clipping the last dimension.

The only training function with more than one non-sample dimension which could possibly receive data with an incorrect dimension ordering is the convolutional model. This model already assumes incoming data has the correct dimensionality, so the clipping assumption of dimension ordering doesn't add any additional danger.

We may consider adding code later to enforce that loader functions return tfdatasets with data of the correct dimensionality.

Refactored public API:
- The `clip` attribute of `PackerConfig` and `ClipConfig` is now `Mapping[Hashable, SliceConfig]`, was `Mapping[Hashable, Mapping[str, SliceConfig]]`.

Significant internal changes:
- refactored clipping functionality to use configuration as `Mapping[Hashable, SliceConfig]`, was `Mapping[Hashable, Mapping[str, SliceConfig]]`.