Small changes to improve PureKerasModel serialization (#1690)
Two functions in the precipitative model are turned into layers, so that later I can look at the configuration of a save-loaded model and compare it to the configuration of starting model to ensure PureKerasModel is serializing correctly.

A get_config method is added to a normalization layer which takes an initialization argument, so it can be used within PureKerasModel. Without this method, keras complains it requires this method to serialize the layer.