offline diags: grab snapshot from batches (#1705)
This ensures that the data used to generate snapshot diagnostics was
also loaded with the given "batches" function. Usually this is
batches_from_mapper, which applies some additional transformations
to the mapper which we would like to happen for the snapshot as well
as the rest of the batched data.