new normalization implementation
I found the multiple-inheritance pattern of normalization.py hard to
follow. It is hard to know how the means and sigmas are computed given
the indirection and names in the class. The design was also tricky to
test since it combined fitting and scaling logic into a single monolith
making it hard to test different types of fitting logic even though
scaling remains unchanged.

* make a new implementation with one normalization class and
several mean and standard deviation fitting functions

this was easier to test than combinatorially many mixin classes

* add an option to compute mean over "all" features rather than
per-feature

* add deprecation warning for old normalization routines

I would delete them now, but they are used by the online-emulation code.