fix `make lock_pip` (#1745)
* Fix make lock_pip on mac

A package `google-cloud-debugging` in the google dataflow image is not
available on mac. We don't use this anyway, so I manually filtered it
out.

Resolves #1721

* Fix dependency locking for fv3fit image

* remove "vcm" dependency from loaders/setup.py

pip-complie cannot lock dependencies that aren't available in a package
repository like pypi

* fix a typo in the makefile

* Remove redundant image requirements.txt from git

these redundant lock files overlapped in responsibility with
contraints.txt. because it is possible to lock just one of these at a
time, they can become out of sync.

Remove from version control and generate them before image construction
to ensure that image builds are always compatible with constraints.txt.

* install pip-copmile in image builds

* add sudo to pip command

* add numpy to circle ci

* fix typo

* fix prognostic-run-gpu build