Prog report: better line plot colors and background (#1768)
Updates the line plots in the prognostic run report to use a color cycle with better visibility as well as gray backgrounds with white grids, and makes black lines for verification. Sources this common color cycle for usage in fv3viz, the prognostic run report, and the logs viewer instead of defining separate cycles in each.