add argo-event to bq webhook
Need to extract argo metadata from k8s and load into big query. There is no
built-in way to do this, so I am using a combination of argo-events and a custom
web server. Argo-events listens for new workflows (the event source) and
forwards (the "sensor") the info as a POST request to the custom web server.

The architecture is potentially over-engineered, but I've been wanting to
explore argo-events for a while now (e.g. as an alternative way to trigger
diagnostics workflows) etc.