Separate base out of prognostic_run image (#1831)
Our caching solution for docker builds sometimes misses, which is very costly in the case of the prognostic run image. This PR separates the costly dependency builds (ESMF, FMS, and NCEPlibs) which we do not modify into their own "base" image which is used as a starting point for the prognostic_run image. This should make it impossible to run into re-compilations of these dependencies in CI.

Instructions for updating this image are included in the header of its Dockerfile.

Significant internal changes:
- defined `prognostic_run_base` and `prognostic_run_base_gpu` images with manually-updated version tag, should be very rarely updated

Resolves #1775