v0.7 base config:  namelist flags set to use climatology  (#1593)
This updates the fv3kube base config to set namelist flags for max snow albedo, vegetation, sea ice, and land sea mask to follow the climatology in the prognostic run, as opposed to the previous behavior of persisting the same values from the initial condition throughout the run.

It also bumps the e2e and prepare-config nudge-to-fine and prognostic tests in fv3net to use the latest base config, and deletes an obsolete reference in fv3kube to a previous default version of the base config. All other references to base configs in fv3net are left at their existing versions.