Offline diagnostics: predict at native and evaluate at C48 (#1756)
Previously the offline diagnostics couldn't operate on test data at non-C48 resolution (at least not on argo). This PR allows the offline diagnostics routine to infer the grid resolution of the test data from its configuration. Offline predictions are made on the native grid of test data; if that resolution is greater than C48 by an integer multiple, the predictions are then coarsened to C48 before being passed for computation of the diagnostics. The default assumption of evaluation at C48 can also now be overridden by passing a path to a grid netcdf at the data resolution, which will be used for evaluation instead (e.g., to locally run the workflow at fully native C384 resolution).

Refactored public API:
- Deleted the `--grid-resolution` argument to `fv3net.diagnostics.offline.compute`. The grid loaded from this argument was mostly being used for adding the grid and establishing the derived mapping in the validation data, which duplicated the same process in `loaders.batches`. It's now assumed that the validation data already computes any such grid-derived variables.
- Renamed the `--grid` argument to `--evaluation-grid` and now use it as a way of evaluating the offline diagnostics at non-C48 resolution if needed, otherwise evaluation at C48 is assumed.
- The diagnostics summary output will not include 3d time-mean predicted variables at evaluation resolutions higher than C384 due to memory constraints, but these are not used in the report generation anyway.