Use smaller version of sample prog run data (#1886)
The prognostic run report test is slow and sometimes fails 
(we think due to OOM issues). This PR updates that test to
use a lighter-weight version of the test data. Specifically
the new data has a time dimension that is about 1/4 as long,
and it has fewer data variables. It should hopefully resolve #1885