Fix bug where segment 2 is run repeatedly
When using the python segemented run api prognostics runs of more than
two segements repeat the second segment indefinitely. This occurs due to
a bug in read_last_segment. fsspec file listings share a write-through
cache that is not updated when GCS is modified outside of fsspec. This
means that listing the artifacts directory will not show the any new
folders uploaded by the gsutil part of append.

The first call to read_last_segment returns None and doesn't call
fsspec.ls. The second call succeeds, and the third call returns the
cached contents of the second call.