Remove tbd parameters gfs radiation driver (#1989)
This PR removes Tbd i/o because the radiation driver module does not use them. 

phy_f3d is only used for flags and values that are set to False or do not apply. 
icsdsw and icsdlw are random generated seeds that were called in mcica_subcol() but they were not used.