Add longitudinal Hovmoller plots to prognostic run report (#1993)
This PR adds longitudinal Hovmoller diagrams to the prognostic run report so that we can get a sense for tropical wave behavior.  It adds an option to set the figure size of 2D plots, which is used to increase the height of the Hovmoller diagrams for longer runs.  An example of what this looks like for a 40-day run can be found [here](https://storage.googleapis.com/vcm-ml-public/argo/tropical-hovmoller-with-verification-no-daily-mean/process_diagnostics.html).

It also adds the option to include area weights when computing zonal or meridional means.