Make copy of inputs before calling transformed function
Otherwise the model is making predictions using the state after
func(), and essentially we are doing both the update from func()
and the model prediction. This definitely applies for the unit
test, not sure if it's relevant when doing a prognostic run and
using the fv3gfs state (versus the mock state in the unit test).