Prognostic run: Simpler radiation stepper factory (#2063)
Adds a cleaner factory method for the radiation stepper, separating out the construction of the prescriber/ML stepper that generates inputs to the radiation stepper, from the construction of the stepper itself. Also in a breaking change for prescribing states, now requires providing a mapping for renaming variables on disk to standard names. 

Refactored public API:
- `runtime.PrescriberConfig.variables` now must be a mapping of variable names on disk to standard names for prescribing (previously was a list of names and no renaming was possible). 

Significant internal changes:
- In the prognostic run loop, the prephysics and radiation input steppers are now generated by a common method, `_get_prescriber_or_ml_stepper`