Add general `setup_environment.sh` script for building the prognostic run environment (#2047)
This PR adds a script that unifies the steps required to build the prognostic run environment on different platforms. The script has breakpoints which allow it to be used in building the prognostic run image without disrupting our existing caching behavior, but otherwise is generally meant to be run all the way through.  It has been tested in Docker in CI and on Gaea manually.

Some documentation for how to use the script can be found in `.environment-scripts/README.md`.  I imagine eventually I may also want to add some documentation for how one might run the model on Gaea as well.

### Requirement changes

- It updates ESMF from 8.0.0 to 8.1.0.  See #2039 for more explanation on the motivation.
- It bumps the version of the fortran repository to include updates to the build configuration files, which makes things easier to unify.

Resolves #2039 
Resolves #2046