Override volume mount for `external/fv3gfs-fortran` in `make enter` rule (#2093)
This is a simple way to address #2080.  If, in addition to the general mount of the user's full `external` directory, we add a specific mount of `/fv3net/external/fv3gfs-fortran`, docker will use the image's `fv3gfs-fortran` directory, but the user's `external` directory for everything else.

In images that don't use fv3gfs-fortran, this will add an empty directory (but previously, in those images we were adding the full user's directory of fv3gfs-fortran, which wasn't really necessary either, so I don't think this is necessarily a step back).  

Resolves #2080