Enable updateable pcolormesh_cube plots (#2102)
The DSL team will soon need to make a video using very high-resolution data. Plotting each frame independently is much slower than running a single pcolormesh and updating its data, but the current interface doesn't allow updating data as it throws away most pcolormesh handles.

This PR refactors the fv3viz pcolormesh_cube code to expose a version of the plotting that retains the handles. This is done intentionally as an "internal" change because the plotting needs for the DSL team are short-term and one-time, but I could make this a public API if reviewers think that's a good idea.

No tests are added.

Significant internal changes:
- Added fv3viz._plot_cube.UpdateablePColormesh