Add Python implementation of blended coarse-graining (#2105)
This PR adds a Python implementation of the blended coarse-graining method outlined in [this document](https://drive.google.com/file/d/1FyLTnR1C5_Ab5Tdbbuhtxm52VC-NroJG/view) by Chris.  This is useful as a way to check the fortran implementation, and if we ever want to use this approach offline.

Added public API:
- `vcm.cubedsphere.coarse_restarts.coarsen_restarts_via_blended_method`
- `vcm.cubedsphere.coarsen.block_edge_coarsen`

These two notebooks show that the implied blending factor-- something you can back out once you have the hybrid coarsened, pressure-level coarsened, and model-level coarsened datasets--for some arbitrary columns looks as we would expect it to:
- [Mass-weighted for most 3D variables (our original default)](https://github.com/ai2cm/explore/blob/master/spencerc/2022-10-31-hybrid-coarsening-verification/2022-10-31-mass-weighted-hybrid-coarsening-test.ipynb)
- [Area-weighted for all 3D variables (newly enabled)](https://github.com/ai2cm/explore/blob/master/spencerc/2022-10-31-hybrid-coarsening-verification/2022-10-31-area-weighted-hybrid-coarsening-test.ipynb)

The pressure-level and model-level components themselves were validated in earlier PRs.  The main new aspects in this PR are the blending, and the option to weight strictly by area.

- [x] Tests added

Coverage reports (updated automatically):
- test_unit: [82%](https://output.circle-artifacts.com/output/job/5b194e5d-8ac1-4c15-b0cb-78266ec61e4e/artifacts/0/tmp/coverage/htmlcov-test_unit/index.html)