combine optimizer settings for cyclegan (#2140)
Currently the CycleGAN network configuration has separate configurations for the generator and discriminator optimizers. This is an issue when running hyperparameter sweeps, as there is no way to link the learning rate of these two optimizers when performing sweeps.

This PR merges the configuration for the two optimizers, which will allow hyperparameter sweeps. The relative learning rates of these two models can still be modified by setting different scale factors on the generator and discriminator loss functions.

Refactored public API:
- CycleGANNetworkConfig now has one `optimizer` setting instead of separate `discriminator_optimizer` and `generator_optimizer` settings