Enable nudging of A-grid winds (#2143)
This PR builds on #2127 to enable nudging to A-grid winds in reference data.  These are the only changes that are required (the transformation to the D-grid should happen automatically now in the runfile and pace-util will read in the A-grid winds from restart files and name them `"eastward_wind"` and `"northward_wind"`).  The plots below show a comparison of the nudging tendencies at an arbitrary vertical level and time in simulations nudged to D-grid and A-grid winds from the same reference dataset.  The D-grid tendencies were converted to A-grid tendencies offline.  

![2023-01-19-u-nudging-snapshot (1)](https://user-images.githubusercontent.com/6628425/213489256-bd34c0c4-3fdb-47ce-97c4-1d970a0560db.png)

![2023-01-19-v-nudging-snapshot (1)](https://user-images.githubusercontent.com/6628425/213489295-f72c788e-9a29-42cc-842a-36a2a77bf88c.png)

One can find more documentation of these experiments in [this directory](https://github.com/ai2cm/explore/tree/master/spencerc/2023-01-18-nudge-to-agrid-winds) and [notebook](https://github.com/ai2cm/explore/blob/master/spencerc/2023-01-18-nudge-to-agrid-winds/2023-01-19-a-grid-vs-d-grid-nudging.ipynb).

- [x] Tests added