vcm: in-cloud <-> gridcell-mean condensate mappings (#2157)
Adds functions to `vcm` to convert between in the standard grid-cell mean condensate mixing ratios and the "in-cloud" equivalent. This exactly follows the conversion done in the GFS radiation scheme to compute effective optical properties. Adds the conversion to "in-cloud" condensate to `DerivedMapping`, and adds the conversions back grid-cell mean as `DataTransform`s.

Added public API:
- `gridcell_to_incloud_condensate`, `incloud_to_gridcell_condensate` added to `vcm`
- `incloud_[water,ice]_mixing_ratio` added to `DerivedMapping`
- `cloud_[water,ice]_mixing_ratio_from_incloud` added as `DataTransform`s

- [X] Tests added