Remove extra instance normalization from first discriminator layer (#2182)
In comparison to the reference Zhu et al. model, our cyclegan includes an extra instance normalization on the first discriminator layer. This leads to the model producing output distributions that are scaled relative to the true distribution, as the discriminator is not able to reason about means and variability.

This PR removes the extra instance normalization from this first layer.

Significant internal changes:
- Removed extra instance normalization from Discriminator in cyclegan and recurrent.