Temporarily roll back base image version.
I will delete the existing images tagged with 1.1.1, which I built
prematurely for this PR.  I will then bump the version in the Makefile
again, rebuild and push the base images, and push the Makefile update
to GitHub so the CI runs fully.

This also re-introduces a few small cleanups after Mike's force-push.