Fix bug in to_url preventing segmented runs using absolute paths (#2209)
This pull request introduces a slight modification to the return value
of `to_url` function. This change is motivated by the fact that the
existing configuration of the function cannot properly deal with the HPC
cluster directory structure. Performing the fv3net-style segmented
simulations on HPC cluster (Stellar) requires locating of the restart
files within the output directory. The output directory is local
absolute path rather then the path on the cloud.

Refactored public API:
- `return protocol + "://" + path`: this return value of the `to_url`
function replaces the existing `return protocol + "://" +
path.lstrip("/")`

- [x] Tests added
The test was added to the `test_fsspec_cloud.py` file.
The predicate (`assert` statement) was modified in the
`test_read_last_segment` function (`test_append.py`).
A new predicate (`assert` statement) was added to the
`test_read_last_segment` function (`test_append.py`). It checks the
existence of the supplied path.

Resolves #2202