Fix array reshaping bug in prediction (#2211)
The reservoir models were working as expected when trained on a single
subdomain, but predictions were way off for the same subdomain when
trained on a larger domain (split into parallel trained subdomains).
Models should give the same result for the same subdomain in these two
cases.

I traced the source of the error to the final readout prediction step,
where `np.reshape` was used improperly. For array reshaping here and in
future inference code I think it would be safest to write testable
functions so that a short example of the desired reshaping can unit
tested. In this short PR I just added a reshaping function that calls a
one-liner of np.reshape for a 2d -> 1d case, but in future PRs the
reshaping is much more complex.
 
- [x] Tests added

Resolves #<github issues> [JIRA-TAG]

Coverage reports (updated automatically):