Use generic transform to reduce dims in reservoir (#2241)
Several changes to hybrid reservoir training with dimensionality
reduction transforms:

- updates the reservoir training function to work with any `Transformer`
class to reduce dimensionality in the inputs (can be either a dense
autoencoder or sklearn transform). Adds abstract base class
`Transformer` which has `encode, decode` methods and `n_latent_dims`
attribute which `Autoencoder` and `SkTransformer` subclass from.
 
- Small change to `SkTransformer` so that concatenating inputs given as
[variable, time, feature] -> [time, variable*feature] occurs in `encode`
rather than `predict`. This is because the training, which provides data
in this format, calls `encode` rather than `predict`.

- changes the hybrid inputs to also use the same dimensionality
reduction transform as the reservoir inputs. Previously, the hybrid
'imperfect state' data was left as is with full vertical dimension, but
this uses a lot of memory and constrains the size of reservoir that can
be used. The linear regression readout training involves saving a
a matrix X_T.X which has (reservoir_state_size +
(hybrid_vars*subdomain_size)) **2 elements, which is updated for each
batch of training data seen. Initializating this matrix takes ~500 Mb
per subdomain for a reservoir of state size 2000 and two full columns of
hybrid input (158 vertical levels per grid point) in a 2x2 subdomain
with overlap of 2. This OOM'ed on a vm with 27 Gb of memory available at
the start of training.