Add mapping adapter for HybridReservoir (#2252)
The current RC model uses subdomain decomposition, feature reduction,
and array reshaping to perform a single prediction. We need a simple
interface to use in the stepper prototype that takes in a dataset-like
mapping and transforms that to the expected input data.

This PR creates an adapter to perform all of the necessary steps to take
in a dataset and adds a regression test. This is a minimal PR that
provides a working adapter, but I expect much of the underlying process
to be refactored away.

- [x] Tests added

Resolves #<github issues> [JIRA-TAG]

Coverage reports (updated automatically):

---------

Co-authored-by: Anna Kwa <annak@allenai.org>