Fix use of ReloadableTransformer in reservior model module (#2264)
The definition of `ReloadableTransformer` was duplicated in
`reservoir.model` to conform with some mypy type constraints. This PR
uses `cast` on the loaded `fv3fit.load` call because the only saved
model for the `autoencoder` argument must be of that type by
construction. This removed the need for that small duplication.