Reservoir warm start (#2327)
With the long-diagnostic runs we now have the ability to pre-synchronize
a model. This PR adds the ability to load models at runtime without
resetting the reservoir state. Additionally, the state is now a field
dumped/loaded with a reservoir model.

Refactored public API:
- The runtime config for reservoirs now includes a `warm_start` option
which will forego state resets on the first increment

- [x] Tests added

Coverage reports (updated automatically):