Update offline diagnostic's use of temporary direcotry (#2342)
The current way of calling temporary directory caused inconsistent
behavior. Most of the time it works fine when the input number of
timesteps is small but crashes for large number of timesteps. I suspect
the temporary directory gets cleaned out somewhat automatically after
`temporary_directory()`. Moving the call to the with context fixes this
inconsistency and we're able to read the batches from the temp dir
regardless of timesteps.

Resolves #<github issues> [JIRA-TAG]

Coverage reports (updated automatically):