Add input mask option for hybrid readout data (#2344)
In the reservoir models, we have options to mask the reservoir inputs
using a specified mask (used for SST runs, currently) but the hybrid
inputs had no such option. This PR takes that same mask and also applies
it to the readout training by multiplication with the post-encoded
spatial gridpoints. It is then saved as a field with the hybrid model
which is loaded and applied to any hybrid input data for prediction.
This allows us to explore how much of the model behavior is tied to
continental atmospheric information vs. the ocean areas where the model
should actually experience the coupling.

- [x] Tests added


Coverage reports (updated automatically):