Hook prognostic_run_shield image into prognostic run argo workflow
I initially implemented this by parametrizing the image, but I've come to
realize that this prevents using kustomize to select a specific image
associated with a commit.  To address this I have switched to using conditional
workflows.  While this revised approach works, it could do with some
de-duplication of definitions.