Remove constraints from GitHub in constraints.txt
In testing, it appears that unnamed constraints don't really have an effect on
what pip installs anyway, so they only seem to cause problems if included in
the constraints.txt file.

We retain dependencies from GitHub in requirements files passed to pip-compile,
as it can resolve dependencies of GitHub-based packages and use that to inform
the versions pinned.