Prognostic run: refactor `main.py` to import wrapper depending on the config (#2362)
This PR is an initial piece broken out of #2350. It adds a `wrapper`
attribute to the `UserConfig` class that is used to determine which
Python wrapper to import within the `main.py` run file. As of this PR,
the only supported option is the default `"fv3gfs.wrapper"`, but in the
abstract this adds a hook for us to ultimately also support
`"shield.wrapper"`.

Added public API:
- A `wrapper` can now be specified within a prognostic run `UserConfig`.
The only currently supported option is the default of
`"fv3gfs.wrapper"`. Because of this default, this PR is fully backwards
compatible (no need to change already-defined configs).

Significant internal changes:
- The Python wrapper is now imported within the `main` function defined
in `main.py` within the prognostic run instead of the top of the module.
- The wrapper is initialized within the `main` function, and thus after
all top-level imports. Initial tests (including with a 2x2 layout in the
cloud) suggest the odd issue that motivated the previous approach may be
resolved. Perhaps this is related to the fact that we no longer use
OpenMPI in the prognostic run image (we use MPICH). This logic had
originally been added in https://github.com/ai2cm/fv3net/pull/706.
- The garbage collection step added in
https://github.com/ai2cm/fv3net/pull/1440 has also been removed, since
it appears it is no longer needed.

- [x] Tests added

Coverage reports (updated automatically):